pyproject.toml
[project]
name = "workshop-aovivo-airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Eduardo Henrique",email = "eduardo.hnsantos@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "apache-airflow (>=3.1.0,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
